requirements.txt
pymongo==4.3.3
python-dotenv==1.0.0
Flask==2.2.3
Flask-Cors==3.0.10
flask_restful
flask_pymongo
